# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger api document for Go Gin Boilerplate.",
        "title": "Go Gin Boilerplate API",
        "contact": {
            "name": "Dorvan",
            "url": "https://github.com/sanjayheaven"
        },
        "version": "v1.0.0"
    },
    "paths": {
        "/examples/createExample": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Example",
                "tags": [
                    "Example"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/examples/deleteExample": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Example",
                "tags": [
                    "Example"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/examples/getExample": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Example",
                "tags": [
                    "Example"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the example id",
                        "name": "exampleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/examples/updateExample": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Example",
                "tags": [
                    "Example"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/public/ping": {
            "get": {
                "description": "Ping",
                "tags": [
                    "Public"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/loginByUsernamePassword": {
            "post": {
                "description": "Login By Username Password",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "username、password",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginByUsernamePasswordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "controllers.LoginByUsernamePasswordRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "default": "123456"
                },
                "username": {
                    "type": "string",
                    "default": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}